// shipeasy/src/app/models/New-payment.ts
export interface CreditDebitNote {
    totalAmount: number;
    tax: any;
    isSelected: any;
    debitDate: any;
    debitToName: string;
    batch_no: string;
    _id: { $oid: string };
    createdOn: string;
    creditdebitnoteId: string;
    creditNoteNo: string;
    createdBy: string;
    createdByUID: string;
    updatedOn: string;
    updatedBy: string;
    tenantId: string;
    invoiceId: string;
    creditDate: { $date: string };
    invoiceType: string;
    remarks: string;
    invoiceNumber: string;
    invoiceParty: string;
    creditToId: string;
    creditToName: string;
    isCredit: boolean;
    amountReceived: number;
    invoiceAmount: string;
    invoiceTaxAmount: string;
    documents: any[];
    paymentTerms: string;
    costItems: CostItem[];
    refundBase: boolean;
    reasonId: string;
    reason: string;
    bankType: string;
    moveNo: number;
    batchId: string;
    vesselName: string;
    voyageNumber: string;
    paymentMode: string;
    payment_ref_no: string;
    currency: string;
    isExport: boolean;
    status: boolean;
    module: string;
    gstType: string;
    __v: number;
  }
  
  interface CostItem {
    tenantId: string;
    enquiryitemId: string;
    enquiryId: string;
    enquiryNumber: string;
    enqDate: { $date: string };
    collectPort: string;
    containerType: string;
    stcQuotationNo: string;
    enqType: string;
    costItemId: string;
    accountBaseCode: string;
    costItemName: string;
    costHeadId: string;
    currency: string;
    exchangeRate: string;
    amount: number;
    baseAmount: string;
    tenantMargin: string;
    tax: Tax[];
    quantity: number;
    rate: number;
    stcAmount: number;
    jmbAmount: number;
    payableAt: string;
    gst: number;
    totalAmount: number;
    chargeTerm: string;
    remarks: string;
    containerNumber: string;
    shippingLine: string;
    taxApplicability: string;
    hsnCode: string;
    isFreight: boolean;
    referenceId: string;
    createdOn: { $date: string };
    updatedOn: { $date: string };
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
    isEnquiryCharge: boolean;
    batchId: string;
    moveNumber: number;
    batchNo: string;
    agentadviceId: string;
    orgId: string;
    vendorId: string;
    vendorName: string;
    exemp: boolean;
    isSelected: boolean;
    currencyShippingLine: string;
    exRateShippingLine: number;
    taxAmount: number;
    gstType: string;
    igst: number;
    cgst: number;
    sgst: number;
    refundAmount: number;
    isRefundSelected: boolean;
    jmb: boolean;
    isRefunds: boolean;
    refundPendingAmount: number;
    _id: { $oid: string };
  }
  
  interface Tax {
    taxAmount: number;
    taxRate: number;
    _id: { $oid: string };
  }
  